Rename path_objective to objectives
